[MISC] Move seqan3::views::to_rank to seqan3/alphabet/views/
