+++ include/seqan3/range/container/bitcompressed_vector.hpp
@@ -21,9 +21,9 @@
 
 #include <seqan3/alphabet/detail/alphabet_proxy.hpp>
 #include <seqan3/alphabet/views/to_char.hpp>
+#include <seqan3/alphabet/views/to_rank.hpp>
 #include <seqan3/core/concept/cereal.hpp>
 #include <seqan3/range/detail/random_access_iterator.hpp>
-#include <seqan3/range/views/to_rank.hpp>
 #include <seqan3/utility/math.hpp>
 #include <seqan3/utility/views/convert.hpp>
