+++ test/unit/range/views/to_rank_test.cpp
@@ -10,8 +10,8 @@
 #include <seqan3/std/ranges>
 
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
+#include <seqan3/alphabet/views/to_rank.hpp>
 #include <seqan3/range/concept.hpp>
-#include <seqan3/range/views/to_rank.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
 
 using seqan3::operator""_dna5;
